% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{buildGeneIndex}
\alias{buildGeneIndex}
\alias{buildGeneIndex}
\alias{buildGeneIndex.SCESet}
\alias{buildGeneIndex,SCESet-method}
\alias{buildGeneIndex}
\title{Build gene index for a dataset}
\usage{
buildGeneIndex(object = NULL, cell_type_column = "cell_type1")

buildGeneIndex.SCESet(object, cell_type_column)

\S4method{buildGeneIndex}{SCESet}(object = NULL,
  cell_type_column = "cell_type1")
}
\arguments{
\item{object}{object SCESet set}

\item{cell_type_column}{column name in the pData slot of the object SCESet 
containing the cell classification information}
}
\value{
a `data.frame` containing calculated gene index
}
\description{
Calculates a fraction of expressed cells per gene per cell type
}
\examples{
library(scater)
pd <- AnnotatedDataFrame(ann)
sceset <- newSCESet(fpkmData = yan, phenoData = pd, logExprsOffset = 1)
sceset <- calculateQCMetrics(sceset)
# use gene names as feature symbols
fData(sceset)$feature_symbol <- featureNames(sceset)
# remove features with duplicated names
sceset <- sceset[!duplicated(fData(sceset)$feature_symbol), ]
gene_index <- buildGeneIndex(sceset)

}
