% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{createReference}
\alias{createReference}
\alias{createReference}
\alias{createReference.SCESet}
\alias{createReference,SingleCellExperiment-method}
\alias{createReference}
\title{Create a precomputed Reference}
\usage{
createReference(reference = NULL, cell_type_column = "cell_type1")

createReference.SCESet(reference, cell_type_column)

\S4method{createReference}{SingleCellExperiment}(reference = NULL,
  cell_type_column = "cell_type1")
}
\arguments{
\item{reference}{reference SingleCellExperiment object}

\item{cell_type_column}{column name in the pData slot of the reference SingleCellExperiment object 
containing the cell classification information}
}
\value{
a `data.frame` containing calculated centroids of the cell types of
the Reference dataset
}
\description{
Calculates centroids of each cell type and merge them into a single table.
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(normcounts = as.matrix(yan)), colData = ann)
# this is needed to calculate dropout rate for feature selection
# important: normcounts have the same zeros as raw counts (fpkm)
counts(sce) <- normcounts(sce)
logcounts(sce) <- log2(normcounts(sce) + 1)
# use gene names as feature symbols
rowData(sce)$feature_symbol <- rownames(sce)
isSpike(sce, 'ERCC') <- grepl('^ERCC-', rownames(sce))
# remove features with duplicated names
sce <- sce[!duplicated(rownames(sce)), ]
sce <- getFeatures(sce)
reference <- createReference(sce[rowData(sce)$scmap_features, ])

}
