% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EuclSqNorm}
\alias{EuclSqNorm}
\title{The Euclidean Squared Norm of each column of a matrix is computed and the whole result is returned as a vector.
Used as part of the approx. calculations of the cosine similarity between the query and the reference.}
\usage{
EuclSqNorm(dat)
}
\arguments{
\item{dat}{A numerical matrix}
}
\description{
The Euclidean Squared Norm of each column of a matrix is computed and the whole result is returned as a vector.
Used as part of the approx. calculations of the cosine similarity between the query and the reference.
}
