% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking.R
\name{fsc3_map_by_similarity}
\alias{fsc3_map_by_similarity}
\title{Bimodality is calculated using dip test: https://en.wikipedia.org/wiki/Multimodal_distribution#General_tests}
\usage{
fsc3_map_by_similarity(object_reference = NULL, object_to_map,
  buckets_ref = NULL, exprs_values = "exprs", similarity = "cosine",
  threshold = 1, suppress_plot = TRUE)
}
\description{
Bimodality is calculated using dip test: https://en.wikipedia.org/wiki/Multimodal_distribution#General_tests
}

