% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{getFeatures.SCESet}
\alias{getFeatures}
\alias{getFeatures,SCESet-method}
\alias{getFeatures.SCESet}
\title{Define a set of genes used for creating cell binary signatures}
\usage{
getFeatures.SCESet(object, n_genes = 100, pct_dropout_min = 20,
  pct_dropout_max = 80, suppress_plot = T)

\S4method{getFeatures}{SCESet}(object, n_genes = 100, pct_dropout_min = 20,
  pct_dropout_max = 80, suppress_plot = T)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{n_genes}{number of the genes to be returned}
}
\value{
an object of 'SCESet' class
}
\description{
The important genes (hyperplanes) are defined as differentially expressed
genes identified using M3Drop (Michaelis-Menten Modelling of Dropouts for scRNASeq,
http://bioconductor.org/packages/M3Drop). A 'data.frame' with the genes
and corresponding p and q values is written to the 'hyperplanes' item of the
'sc3' slot of the input object.
}

