% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{getSignatures.SCESet}
\alias{getSignatures}
\alias{getSignatures,SCESet-method}
\alias{getSignatures.SCESet}
\title{Create a binary signature for each cell}
\usage{
getSignatures.SCESet(object, exprs_values = "exprs")

\S4method{getSignatures}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{exprs_values}{}
}
\value{
an object of 'SCESet' class
}
\description{
Create a binary signature for each cell
}

