% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{mapData}
\alias{mapData}
\alias{mapData.SCESet}
\alias{mapData,SCESet-method}
\alias{mapData}
\title{scmap main function}
\usage{
mapData.SCESet(object_map, object_ref, class_col, class_ref, method, similarity,
  threshold, ml_threshold, scale_exprs, suppress_plot)

\S4method{mapData}{SCESet}(object_map = NULL, object_ref = NULL,
  class_col = "cell_type1", class_ref = NULL, method = "scmap",
  similarity = "cosine", threshold = 0.5, ml_threshold = 0.5,
  scale_exprs = FALSE, suppress_plot = TRUE)
}
\arguments{
\item{object_map}{SCESet to map}

\item{object_ref}{reference SCESet set}

\item{class_col}{column name in the pData slot of the reference SCESet containing the cell classification information}

\item{class_ref}{reference cell buckets}

\item{similarity}{similarity measure}

\item{threshold}{threshold on similarity}

\item{ml_threshold}{threshold on probability for SVM}

\item{scale_exprs}{scale expression or not?}

\item{suppress_plot}{defines whether to suppress an output plot}
}
\description{
Mapping of one dataset to another
}
