% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfinemap_fns.R
\name{mult_search}
\alias{mult_search}
\title{For each cell in a query dataset, we search for the nearest neighbours by cosine distance
within a collection of reference datasets.}
\usage{
mult_search(list_dat, query_dat, w)
}
\arguments{
\item{list_dat}{list of index objects each coming from the output of scf_index}

\item{query_dat}{an object of \code{\link[SingleCellExperiment]{SingleCellExperiment}} class}

\item{w}{a positive integer specifying the number of nearest neighbours to find}
}
\value{
a list of 3 objects: 
1) a matrix with the closest w neighbours by cell number of each query cell stored by column
2) a matrix of integers giving the reference datasets from which the above cells came from
3) a matrix with the cosine similarities corresponding to each of the nearest neighbours
}
\description{
For each cell in a query dataset, we search for the nearest neighbours by cosine distance
within a collection of reference datasets.
}
