% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NNmult}
\alias{NNmult}
\title{Performs the same operations as NNfirst, but for subsequent reference datasets.
The current best cells and distances are given as inputs and are updated as the reference is searched.}
\usage{
NNmult(w, k, subcentroids, subclusters, query_chunks, M, SqNorm,
  best_cells_so_far, best_distances_so_far, dataset_inds, dat_num)
}
\arguments{
\item{w}{An integer specifying the number of nearest neighbours}

\item{k}{An integer specifying the number of subcentroids for each product quantization chunk}

\item{subcentroids}{A list of matrices containing the subcentroids of each chunk.}

\item{subclusters}{A matrix containing the subcentroid assignments of each reference cell. See scf_index.}

\item{query_chunks}{A list of matrices containing the chunks of the query dataset after it has been split
according to the product quantization method}

\item{M}{An integer specifying the number of chunks}

\item{SqNorm}{A numerical vector containing the Euclidean Squared Norm of each query cell.}

\item{best_cells_so_far}{A integer matrix containing the cell indices of the nearest neighbours so far.}

\item{best_distances_so_far}{A numerical matrix containing the approx. cosine similarities of the 
nearest neighbours so far.}

\item{dataset_inds}{An integer matrix containing the dataset indices of the nearest neighbours so far.}

\item{dat_num}{The reference dataset number in consideration (i.e. the position of the reference dataset
in the list of references).}
}
\description{
Performs the same operations as NNfirst, but for subsequent reference datasets.
The current best cells and distances are given as inputs and are updated as the reference is searched.
}
