% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{plotFeatures}
\alias{plotFeatures}
\title{Plot feature distibution and highlight the most informative features.}
\usage{
plotFeatures(object, n_features = 500)
}
\arguments{
\item{object}{an object of \code{\link[scater]{SCESet}} class}

\item{n_features}{number of the features to be selected.}
}
\description{
Plots log(dropout_rate) versus log2(mean_expression) for all genes. 
Selected features are highlight with the red colour.
}
