% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{projectData}
\alias{projectData}
\alias{projectData.SCESet}
\alias{projectData,SCESet-method}
\alias{projectData}
\title{scmap main function}
\usage{
projectData.SCESet(projection, reference, cell_type_column, method, threshold)

\S4method{projectData}{SCESet}(projection = NULL, reference = NULL,
  cell_type_column = "cell_type1", method = "scmap", threshold = 0.7)
}
\arguments{
\item{projection}{SCESet to project}

\item{reference}{reference SCESet set}

\item{cell_type_column}{column name in the pData slot of the reference SCESet containing the cell classification information}

\item{method}{which method to use}

\item{threshold}{threshold on similarity (or probability for SVM and RF)}
}
\description{
Mapping of one dataset to another
}
