% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking.R
\name{random_forest}
\alias{random_forest}
\title{Run random forest prediction}
\usage{
random_forest(train, study, ntree = 50)
}
\arguments{
\item{train}{training data.frame with colnames, corresponding to training labels}

\item{study}{study data.frame}

\item{ntree}{number of trees to be used in random forest}
}
\value{
classification of the study dataset
}
\description{
Create a random forest calssifier based on a training dataset (\code{train}) and then
classify a study dataset (\code{study}) using the classifier.
}

