% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfinemap_fns.R
\name{scf_celltype}
\alias{scf_celltype}
\title{Approximate k-NN cell-type classification using scfinemap}
\usage{
scf_celltype(ref, query_dat, ref_index = NULL, w = 3, thres = 0.5)
}
\arguments{
\item{ref}{an object of \code{\link[SingleCellExperiment]{SingleCellExperiment}} class}

\item{query_dat}{an object of \code{\link[SingleCellExperiment]{SingleCellExperiment}} class}

\item{ref_index}{the output of scf_index with ref as its input.}
}
\value{
the predicted cell-type labels of the query dataset
}
\description{
Each cell in the query dataset is assigned a cell-type if the similarity between its
nearest neighbour exceeds a threshold AND its w nearest neighbours have the 
same cell-type.
}
