% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfinemap_fns.R
\name{scf_index}
\alias{scf_index}
\title{Create an index for a dataset to enable fast approximate nearest neighbour search}
\usage{
scf_index(dat, M = 100, k = NULL)
}
\arguments{
\item{dat}{an object of \code{\link[SingleCellExperiment]{SingleCellExperiment}} class}

\item{M}{number of chunks into which the expr matrix is split}

\item{k}{number of clusters per group for k-means clustering}
}
\value{
a list of four objects: 1) a list of matrices containing the subcentroids of each group
2) a matrix containing the subclusters for each cell for each group
3) the value of M
4) the value of k
}
\description{
The method is based on product quantization for the cosine distance.
Split the training data into M identically sized chunks by genes.
Use k-means to find k subcentroids for each group.
Assign cluster numbers to each member of the dataset.
}
