% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{scmap}
\alias{scmap}
\title{Bimodality is calculated using dip test: https://en.wikipedia.org/wiki/Multimodal_distribution#General_tests}
\usage{
scmap(object_reference = NULL, object_to_map, buckets_ref = NULL,
  exprs_values = "exprs", similarity = "cosine", threshold = 1,
  suppress_plot = TRUE)
}
\arguments{
\item{object_reference}{reference SCESet set}

\item{object_to_map}{SCESet to map}

\item{buckets_ref}{reference cell buckets}

\item{exprs_values}{expression values to use}

\item{similarity}{similarity measure}

\item{threshold}{threshold on similarity}

\item{suppress_plot}{defines whether to suppress an output plot}
}
\description{
Bimodality is calculated using dip test: https://en.wikipedia.org/wiki/Multimodal_distribution#General_tests
}
