% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{setFeatures.SCESet}
\alias{setFeatures}
\alias{setFeatures,SCESet-method}
\alias{setFeatures.SCESet}
\title{Title}
\usage{
setFeatures.SCESet(object, features)

\S4method{setFeatures}{SCESet}(object, features)
}
\arguments{
\item{features}{a vector of feature names}
}
\description{
Title
}

