% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{buildGeneIndex}
\alias{buildGeneIndex}
\alias{buildGeneIndex}
\alias{buildGeneIndex.SCESet}
\alias{buildGeneIndex,SingleCellExperiment-method}
\alias{buildGeneIndex}
\title{Build gene index for a dataset}
\usage{
buildGeneIndex(object = NULL, cell_type_column = "cell_type1")

buildGeneIndex.SCESet(object, cell_type_column)

\S4method{buildGeneIndex}{SingleCellExperiment}(object = NULL,
  cell_type_column = "cell_type1")
}
\arguments{
\item{object}{object of SingleCellExperiment class}

\item{cell_type_column}{column name in the colData slot of the object SingleCellExperiment 
containing the cell classification information}
}
\value{
a `data.frame` containing calculated gene index
}
\description{
Calculates a fraction of expressed cells per gene per cell type
}
\examples{
library(SingleCellExperiment)
sce <- SingleCellExperiment(assays = list(normcounts = as.matrix(yan)), colData = ann)
# this is needed to calculate dropout rate for feature selection
# important: normcounts have the same zeros as raw counts (fpkm)
counts(sce) <- normcounts(sce)
logcounts(sce) <- log2(normcounts(sce) + 1)
# use gene names as feature symbols
rowData(sce)$feature_symbol <- rownames(sce)
isSpike(sce, "ERCC") <- grepl("^ERCC-", rownames(sce))
# remove features with duplicated names
sce <- sce[!duplicated(rownames(sce)), ]
gene_index <- buildGeneIndex(sce)

}
