% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{createReference}
\alias{createReference}
\title{Create a precomputed Reference}
\usage{
createReference(reference, cell_type_column = "cell_type1")
}
\arguments{
\item{reference}{reference SCESet set}

\item{cell_type_column}{column name in the pData slot of the reference SCESet containing the cell classification information}
}
\description{
Calculates centroids of each cell type and merge them into a single table.
}
