% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{createReference}
\alias{createReference}
\alias{createReference}
\alias{createReference.SCESet}
\alias{createReference,SCESet-method}
\alias{createReference}
\title{Create a precomputed Reference}
\usage{
createReference(reference = NULL, cell_type_column = "cell_type1")

createReference.SCESet(reference, cell_type_column)

\S4method{createReference}{SCESet}(reference = NULL,
  cell_type_column = "cell_type1")
}
\arguments{
\item{reference}{reference SCESet set}

\item{cell_type_column}{column name in the pData slot of the reference SCESet 
containing the cell classification information}
}
\value{
a `data.frame` containing calculated centroids of the cell types of
the Reference dataset
}
\description{
Calculates centroids of each cell type and merge them into a single table.
}
\examples{
library(scater)
pd <- AnnotatedDataFrame(ann)
sceset <- newSCESet(fpkmData = yan, phenoData = pd, logExprsOffset = 1)
sceset <- calculateQCMetrics(sceset)
# use gene names as feature symbols
fData(sceset)$feature_symbol <- featureNames(sceset)
# remove features with duplicated names
sceset <- sceset[!duplicated(fData(sceset)$feature_symbol), ]
sceset <- getFeatures(sceset)
reference <- createReference(sceset[fData(sceset)$scmap_features, ])

}
