% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{getFeatures}
\alias{getFeatures}
\alias{getFeatures}
\alias{getFeatures.SCESet}
\alias{getFeatures,SCESet-method}
\alias{getFeatures}
\title{Select the most informative features (genes/transcripts) for projection}
\usage{
getFeatures(object, n_features = 500, suppress_plot = TRUE)

getFeatures.SCESet(object, n_features, suppress_plot)

\S4method{getFeatures}{SCESet}(object, n_features = 500,
  suppress_plot = TRUE)
}
\arguments{
\item{object}{an object of \code{\link[scater]{SCESet}} class}

\item{n_features}{number of the features to be selected}

\item{suppress_plot}{boolean parameter, which defines whether to plot 
log(expression) versus log(dropout) distribution for all genes.
Selected features are highlighted with the red colour.}
}
\value{
an object of \code{\link[scater]{SCESet}} class with a new column in 
\code{featureData} slot which is called \code{scmap_features}. It can be accessed
by using \code{fData(object)$scmap_features}.
}
\description{
This is a modification of the M3Drop method. Instead of fitting a 
Michaelis-Menten model to the log expression-dropout relation, we fit a 
linear model. Namely, the linear model is build on the log(expression) versus 
log(dropout) distribution. After fitting a linear model important features are
 selected as the top N residuals of the linear model.
}
\details{
Please note that \code{object@featureData@data$feature_symbol} column must be 
present in the input object and should not contain any duplicated feature names. 
This column defines feature names used during projection. Feature symbols 
in the reference dataset must correpond to the feature symbols
in the projection dataset, otherwise the mapping will not work!
}
\examples{
library(scater)
pd <- AnnotatedDataFrame(ann)
sceset <- newSCESet(fpkmData = yan, phenoData = pd, logExprsOffset = 1)
sceset <- calculateQCMetrics(sceset)
# use gene names as feature symbols
fData(sceset)$feature_symbol <- featureNames(sceset)
# remove features with duplicated names
sceset <- sceset[!duplicated(fData(sceset)$feature_symbol), ]
sceset <- getFeatures(sceset)

}
