% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{mapData}
\alias{mapData}
\alias{mapData.SCESet}
\title{scmap main function}
\usage{
mapData.SCESet(object_map, object_ref, class_col, class_ref, threshold,
  suppress_plot)
}
\arguments{
\item{object_map}{SCESet to map}

\item{object_ref}{reference SCESet set}

\item{class_col}{column name in the pData slot of the reference SCESet containing the cell classification information}

\item{class_ref}{reference cell buckets}

\item{threshold}{threshold on similarity}

\item{suppress_plot}{defines whether to suppress an output plot}
}
\description{
scmap main function
}
