% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{plotFeatures}
\alias{plotFeatures}
\title{Plot feature selection plot}
\usage{
plotFeatures(object, n_features = 500)
}
\arguments{
\item{object}{an object of \code{\link[scater]{SCESet}} class}

\item{n_features}{number of features to select}
}
\value{
a ggplot object to plot the feature selection plot
}
\description{
Plot feature selection plot
}
