% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{projectData}
\alias{projectData}
\alias{projectData.SCESet}
\alias{projectData,SCESet-method}
\alias{projectData}
\title{scmap main function}
\usage{
projectData.SCESet(map, object_ref, class_col, class_ref, method, threshold)

\S4method{projectData}{SCESet}(map = NULL, object_ref = NULL,
  class_col = "cell_type1", class_ref = NULL, method = "scmap",
  threshold = 0.7)
}
\arguments{
\item{object_ref}{reference SCESet set}

\item{class_col}{column name in the pData slot of the reference SCESet containing the cell classification information}

\item{class_ref}{reference cell buckets}

\item{method}{which method to use}

\item{threshold}{threshold on similarity (or probability for SVM and RF)}

\item{object_map}{SCESet to map}
}
\description{
Mapping of one dataset to another
}
