% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{projectData}
\alias{projectData}
\alias{projectData}
\alias{projectData.SCESet}
\alias{projectData,SCESet-method}
\alias{projectData}
\title{scmap main function}
\usage{
projectData(projection = NULL, reference = NULL,
  cell_type_column = "cell_type1", method = "scmap", threshold = 0.7)

projectData.SCESet(projection, reference, cell_type_column, method, threshold)

\S4method{projectData}{SCESet}(projection = NULL, reference = NULL,
  cell_type_column = "cell_type1", method = "scmap", threshold = 0.7)
}
\arguments{
\item{projection}{SCESet to project}

\item{reference}{reference SCESet set}

\item{cell_type_column}{column name in the pData slot of the reference SCESet containing the cell classification information}

\item{method}{which method to use}

\item{threshold}{threshold on similarity (or probability for SVM and RF)}
}
\value{
Projection SCESet object with labels calculated by `scmap` stored in 
the `scmap_labels` column of the `phenoData` slot.
}
\description{
Projection of one dataset to another
}
\examples{
library(scater)
pd <- AnnotatedDataFrame(ann)
sceset <- newSCESet(fpkmData = yan, phenoData = pd, logExprsOffset = 1)
sceset <- calculateQCMetrics(sceset)
# use gene names as feature symbols
fData(sceset)$feature_symbol <- featureNames(sceset)
# remove features with duplicated names
sceset <- sceset[!duplicated(fData(sceset)$feature_symbol), ]
sceset <- getFeatures(sceset)
sceset <- projectData(projection = sceset, reference = sceset)

}
