% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/CoreMethods.R
\docType{methods}
\name{setFeatures}
\alias{setFeatures}
\alias{setFeatures}
\alias{setFeatures.SCESet}
\alias{setFeatures,SCESet-method}
\alias{setFeatures}
\title{Set the most important features (genes/transcripts) for mapping}
\usage{
setFeatures(object, features = NULL)

setFeatures.SCESet(object, features)

\S4method{setFeatures}{SCESet}(object, features = NULL)
}
\arguments{
\item{object}{an object of \code{\link[scater]{SCESet}} class}

\item{features}{a character vector of feature names}
}
\value{
an object of \code{\link[scater]{SCESet}} class with a new column in 
\code{featureData} slot which is called \code{scmap_features}. It can be accessible
either by \code{fData(object)$scmap_features} or by \code{object@featureData@data$scmap_features}
}
\description{
This method set the features to be used for mapping.
}
\details{
Please note that \code{object@featureData@data$feature_symbol} column must be 
present in the input object. This column defines feature names used during mapping
Feature symbols in the reference dataset must correpond to the feature symbols
in the mapping dataset, otherwise the mapping will not work!
}
\examples{
library(scater)
pd <- AnnotatedDataFrame(ann)
sceset <- newSCESet(fpkmData = yan, phenoData = pd, logExprsOffset = 1)
sceset <- calculateQCMetrics(sceset)
# use gene names as feature symbols
fData(sceset)$feature_symbol <- featureNames(sceset)
# remove features with duplicated names
sceset <- sceset[!duplicated(fData(sceset)$feature_symbol), ]
sceset <- setFeatures(sceset, c('MMP2', 'ZHX3'))

}
